move upload logic
