+++ package.json
@@ -26,6 +26,7 @@
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.14",
     "@mui/x-charts": "^7.1.1",
+    "axios": "^1.6.8",
     "browser-image-compression": "^2.0.2",
     "clsx": "^2.0.0",
     "ethers": "^6.11.1",
